requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0
statsmodels>=0.12.0
prophet>=1.0.0
torch>=1.8.0  # For LSTM if using
streamlit>=0.85.0
pyyaml>=5.4.0
tqdm>=4.61.0